Use long instead of int type

+++ fluxc/src/main/java/org/wordpress/android/fluxc/model/blaze/BlazeAdForecast.kt
@@ -1,6 +1,6 @@
 package org.wordpress.android.fluxc.model.blaze
 
 data class BlazeAdForecast(
-    val minImpressions: Int,
-    val maxImpressions: Int,
+    val minImpressions: Long,
+    val maxImpressions: Long,
 )
